+++ install_wordpress_server.sh
@@ -6,11 +6,11 @@
 
 # Edit variables for Plesk pre-configuration
 
-hostname='cp.domain.tst'
-email='admin@test.tst'
-passwd='CookBook123'
-name='admin'
-company='Plesk Sample Company'
+hostname='domain.tld'
+email='root@localhost'
+passwd='ChangeMe'
+name='Name'
+company='Plesk'
 phone='123-123-1234'
 address='123_street'
 city='NY'
@@ -23,6 +23,7 @@ ip_type=shared
 # Plesk Activation Code - provide proper license for initialization, it will be replaced after cloning
 # leave as null if not providing key
 activation_key=
+
 
 # Plesk UI View - can be set to Service Provider View (spv) or Power User View (puv)
 plesk_ui=puv
@@ -69,7 +70,7 @@ echo
 # Install Plesk with Required Components
 
 echo "Starting Plesk Installation"
-./plesk-installer install plesk --preset Full --with panel bind fail2ban l10n pmm mysqlgroup roundcube kav spamassassin selinux postfix dovecot proftpd awstats modsecurity mod_fcgid webservers php7.1 php5.6 config-troubleshooter psa-firewall cloudflare heavy-metal-skin wp-toolkit security-advisor letsencrypt
+./plesk-installer install plesk --preset Full --with panel bind fail2ban l10n pmm mysqlgroup roundcube kav spamassassin postfix dovecot proftpd awstats modsecurity mod_fcgid webservers php7.1 php5.6 config-troubleshooter psa-firewall heavy-metal-skin wp-toolkit letsencrypt >> /tmp/plsklog
 echo
 echo
 
@@ -117,6 +118,7 @@ iptables -I INPUT -p tcp --dport 995 -j ACCEPT
 iptables -I INPUT -p tcp --dport 8443 -j ACCEPT
 iptables -I INPUT -p tcp --dport 8447 -j ACCEPT
 iptables -I INPUT -p tcp --dport 8880 -j ACCEPT
+
 echo
 
 # Enable Modsecurity
@@ -159,21 +161,32 @@ fi
 # https://docs.plesk.com/en-US/onyx/cli-linux/using-command-line-utilities/extension-extensions.71031/
 
 echo "Installing Requested Plesk Extensions"
-echo "Installing Route 53"
-plesk bin extension --install-url https://ext.plesk.com/packages/ed1860ee-45c5-4e2b-b6b7-44e5da69dca5-route53/download
+echo "Installing SEO Toolkit"
+plesk bin extension --install-url https://ext.plesk.com/packages/2ae9cd0b-bc5c-4464-a12d-bd882c651392-xovi/download
 echo
-echo "Installing Security Advisor"
-plesk bin extension --install-url https://ext.plesk.com/packages/6bcc01cf-d7bb-4e6a-9db8-dd1826dcad8f-security-advisor/download
+echo "Installing Advisor"
+plesk bin extension --install-url https://ext.plesk.com/packages/bbf16bc7-094e-4cb3-8b9c-32066fc66561-advisor/download
+echo
+echo "Installing BoldGrid"
+plesk bin extension --install-url https://ext.plesk.com/packages/e4736f87-ba7e-4601-a403-7c82682ef07d-boldgrid/download
+echo
+echo "Installing Backup to Cloud Pro"
+plesk bin extension --install-url https://ext.plesk.com/packages/9f3b75b3-d04d-44fe-a8fa-7e2b1635c2e1-dropbox-backup/download
+plesk bin extension --install-url https://ext.plesk.com/packages/52fd6315-22a4-48b8-959d-b2f1fd737d11-google-drive-backup/download
+plesk bin extension --install-url https://ext.plesk.com/packages/8762049b-870e-47cb-ba14-9f055b99b508-s3-backup/download
+plesk bin extension --install-url https://ext.plesk.com/packages/a8e5ad9c-a254-4bcf-8ae4-5440f13a88ad-one-drive-backup/download
+echo
+echo "Installing Revisium Antivirus for Websites"
+plesk bin extension --install-url https://ext.plesk.com/packages/b71916cf-614e-4b11-9644-a5fe82060aaf-revisium-antivirus/download
 echo
 echo "Installing Google Pagespeed Insights"
 plesk bin extension --install-url https://ext.plesk.com/packages/3d2639e6-64a9-43fe-a990-c873b6b3ec66-pagespeed-insights/download
 echo
-echo "Installing Addendio - WordPress Plugin and Themes"
-plesk bin extension --install-url https://ext.plesk.com/packages/250589ff-8081-4c30-b9ca-66539e025c27-addendio-wordpress/download
+echo "Installing Uptime Robot"
+plesk bin extension --install-url https://ext.plesk.com/packages/7d37cfde-f133-4085-91ea-d5399862321b-uptime-robot/download
 echo
-echo "Installing Datagrid VCTR reliability and vulnerability scanner"
-plesk bin extension --install-url https://ext.plesk.com/packages/e757450e-40a5-44e5-a35d-8c4c50671019-dgri/download
-echo
+echo "Installing Sucuri Site Scanner"
+plesk bin extension --install-url https://ext.plesk.com/packages/2d5b423b-9104-40f2-9286-a75a6debd43f-sucuri-scanner/download
 echo "Installing LetsEncrypt"
 plesk bin extension --install-url https://ext.plesk.com/packages/f6847e61-33a7-4104-8dc9-d26a0183a8dd-letsencrypt/download
 echo
@@ -194,7 +207,7 @@ if [ "$clone" = "on" ]; then
 	plesk bin cloning --update -prepare-public-image true
 	echo "Plesk initialization will be wiped on next boot. Ready for Cloning."
 fi
-
+plesk bin settings --set solution_type="wordpress"
 echo
 echo "Your Plesk WordPress Server image is complete."
 echo "Thank you for using the WordPress Server Cookbook"