+++ CHANGES.md
@@ -1,3 +1,13 @@
+# 1.1
+* [*] Updated the cookbook for the latest version of the WordPress Server
+* [*] Replaced the Security Advisor with our new Advisor Extension
+* [+] Added Extension installation for SEO Toolkit, Advisor, Uptime Robot and many more 
+* [+] Added Solution Type for reporting purposes
+* [-] Removed extensions no longer available in WordPress Server like Opsani or Addendino
+* [-] Removed Route 53 Extension
+
+
+
 # 1.0.1
 
 * [*] Updated link to the latest Welcome WordPress extension package (version 1.0.2)